# Decrease impact of gc on benchmark performance (fixes #185)