@patrickdoc 2018-02-25
Ah, tragically, `performMinorGc` wasn't implemented/exported until base 4.7.0. I've added the CPP to let it build, but it's not the most elegant. Let me know if you would like something different.