@RyanGlScott 2018-02-26
It looks like the CI is failing due https://github.com/snoyberg/conduit/issues/353. Let me see if that's easy to fix...