@RyanGlScott 2018-02-26
Might you be willing to rebase on top of `master` again? I believe https://github.com/bos/criterion/commit/bd5898ba4f705ee228e1d6e03982346f34124b47 should fix the previous build failures.