@patrickdoc 2018-02-26
Well you now know that I don't know how to use ffi correctly... One last fix should do it.